move eq,peq utilities func to a standalone file
