+++ SofaKernel/modules/SofaBaseMechanics/src/SofaBaseMechanics/IdentityMapping.h
@@ -142,9 +142,6 @@ extern template class SOFA_SOFABASEMECHANICS_API IdentityMapping< defaulttype::R
 extern template class SOFA_SOFABASEMECHANICS_API IdentityMapping< defaulttype::Rigid3Types, defaulttype::Vec3dTypes >;
 extern template class SOFA_SOFABASEMECHANICS_API IdentityMapping< defaulttype::Rigid2Types, defaulttype::Vec2Types >;
 
-
-
-
 #endif
 
 } // namespace sofa::component::mapping
@@ -153,131 +150,97 @@ extern template class SOFA_SOFABASEMECHANICS_API IdentityMapping< defaulttype::R
 
 namespace sofa::helper
 {
-    // should certainly be somewhere else
-    // but at least it is accessible to other components
 
+#define SOFA_ATTRIBUTE_DISABLED__HELPER_EQ() \
+    SOFA_ATTRIBUTE_DISABLED( \
+        "v21.06 (PR#21XX)", "v21.06", "You need to include <sofa/core/MappingHelper.h> and use sofa::core::eq(...) instead)")
+
+#define SOFA_ATTRIBUTE_DISABLED__HELPER_PEQ() \
+    SOFA_ATTRIBUTE_DISABLED( \
+        "v21.06 (PR#21XX)", "v21.06", "You need to include <sofa/core/MappingHelper.h> and use sofa::core::peq(...) instead)")
+
+    // Those static functions have been moved in MappingHelper.h
+    template<class T1, class T2>
+    SOFA_ATTRIBUTE_DISABLED__HELPER_EQ()
+    static inline void eq(T1& dest, const T2& src) = delete;
 
     template<class T1, class T2>
-    static inline void eq(T1& dest, const T2& src)
-    {
-        dest = src;
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_PEQ()
+    static inline void peq(T1& dest, const T2& src) = delete;
 
-    template<class T1, class T2>
-    static inline void peq(T1& dest, const T2& src)
-    {
-        dest += src;
-    }
-
-    // float <-> double (to remove warnings)
+    SOFA_ATTRIBUTE_DISABLED__HELPER_EQ()
+    static inline void eq(float& dest, const double& src) = delete;
 
     //template<>
-    static inline void eq(float& dest, const double& src)
-    {
-        dest = (float)src;
-    }
-
-    //template<>
-    static inline void peq(float& dest, const double& src)
-    {
-        dest += (float)src;
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_PEQ()
+    static inline void peq(float& dest, const double& src) = delete;
 
     // Vec <-> Vec
 
     template<Size N1, Size N2, class T1, class T2>
-    static inline void eq(defaulttype::Vec<N1,T1>& dest, const defaulttype::Vec<N2,T2>& src)
-    {
-        dest = src;
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_EQ()
+    static inline void eq(defaulttype::Vec<N1, T1>& dest, const defaulttype::Vec<N2, T2>& src) = delete;
 
     template<Size N1, Size N2, class T1, class T2>
-    static inline void peq(defaulttype::Vec<N1,T1>& dest, const defaulttype::Vec<N2,T2>& src)
-    {
-        for (Size i=0; i<(N1>N2?N2:N1); i++)
-            dest[i] += (T1)src[i];
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_PEQ()
+    static inline void peq(defaulttype::Vec<N1, T1>& dest, const defaulttype::Vec<N2, T2>& src) = delete;
 
     // RigidDeriv <-> RigidDeriv
 
     template<Size N, class T1, class T2>
-    static inline void eq(defaulttype::RigidDeriv<N,T1>& dest, const defaulttype::RigidDeriv<N,T2>& src)
-    {
-        dest.getVCenter() = src.getVCenter();
-        dest.getVOrientation() = (typename defaulttype::RigidDeriv<N,T1>::Rot)src.getVOrientation();
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_EQ()
+    static inline void eq(defaulttype::RigidDeriv<N, T1>& dest, const defaulttype::RigidDeriv<N, T2>& src) = delete;
 
     template<Size N, class T1, class T2>
-    static inline void peq(defaulttype::RigidDeriv<N,T1>& dest, const defaulttype::RigidDeriv<N,T2>& src)
-    {
-        dest.getVCenter() += src.getVCenter();
-        dest.getVOrientation() += (typename defaulttype::RigidDeriv<N,T1>::Rot)src.getVOrientation();
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_PEQ()
+    static inline void peq(defaulttype::RigidDeriv<N, T1>& dest, const defaulttype::RigidDeriv<N, T2>& src) = delete;
 
     // RigidCoord <-> RigidCoord
 
     template<Size N, class T1, class T2>
-    static inline void eq(defaulttype::RigidCoord<N,T1>& dest, const defaulttype::RigidCoord<N,T2>& src)
-    {
-        dest.getCenter() = src.getCenter();
-        dest.getOrientation() = (typename defaulttype::RigidCoord<N,T1>::Rot)src.getOrientation();
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_EQ()
+    static inline void eq(defaulttype::RigidCoord<N, T1>& dest, const defaulttype::RigidCoord<N, T2>& src) = delete;
 
     template<Size N, class T1, class T2>
-    static inline void peq(defaulttype::RigidCoord<N,T1>& dest, const defaulttype::RigidCoord<N,T2>& src)
-    {
-        dest.getCenter() += src.getCenter();
-        dest.getOrientation() += src.getOrientation();
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_PEQ()
+    static inline void peq(defaulttype::RigidCoord<N, T1>& dest, const defaulttype::RigidCoord<N, T2>& src) = delete;
 
     // RigidDeriv <-> Vec
 
     template<Size N, class T1, class T2>
-    static inline void eq(defaulttype::Vec<N,T1>& dest, const defaulttype::RigidDeriv<N,T2>& src)
-    {
-        dest = src.getVCenter();
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_EQ()
+    static inline void eq(defaulttype::Vec<N, T1>& dest, const defaulttype::RigidDeriv<N, T2>& src) = delete;
 
     template<Size N, class T1, class T2>
-    static inline void peq(defaulttype::Vec<N,T1>& dest, const defaulttype::RigidDeriv<N,T2>& src)
-    {
-        dest += src.getVCenter();
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_PEQ()
+    static inline void peq(defaulttype::Vec<N, T1>& dest, const defaulttype::RigidDeriv<N, T2>& src) = delete;
 
     template<Size N, class T1, class T2>
-    static inline void eq(defaulttype::RigidDeriv<N,T1>& dest, const defaulttype::Vec<N,T2>& src)
-    {
-        dest.getVCenter() = src;
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_EQ()
+    static inline void eq(defaulttype::RigidDeriv<N,T1>& dest, const defaulttype::Vec<N,T2>& src) = delete;
 
     template<Size N, class T1, class T2>
-    static inline void peq(defaulttype::RigidDeriv<N,T1>& dest, const defaulttype::Vec<N,T2>& src)
-    {
-        dest.getVCenter() += src;
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_PEQ()
+    static inline void peq(defaulttype::RigidDeriv<N,T1>& dest, const defaulttype::Vec<N,T2>& src) = delete;
 
     // RigidCoord <-> Vec
     template<Size N, class T1, class T2>
-    static inline void eq(defaulttype::Vec<N,T1>& dest, const defaulttype::RigidCoord<N,T2>& src)
-    {
-        dest = src.getCenter();
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_EQ()
+    static inline void eq(defaulttype::Vec<N,T1>& dest, const defaulttype::RigidCoord<N,T2>& src) = delete;
 
     template<Size N, class T1, class T2>
-    static inline void peq(defaulttype::Vec<N,T1>& dest, const defaulttype::RigidCoord<N,T2>& src)
-    {
-        dest += src.getCenter();
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_PEQ()
+    static inline void peq(defaulttype::Vec<N,T1>& dest, const defaulttype::RigidCoord<N,T2>& src) = delete;
 
     template<Size N, class T1, class T2>
-    static inline void eq(defaulttype::RigidCoord<N,T1>& dest, const defaulttype::Vec<N,T2>& src)
-    {
-        dest.getCenter() = src;
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_EQ()
+    static inline void eq(defaulttype::RigidCoord<N,T1>& dest, const defaulttype::Vec<N,T2>& src) = delete;
 
     template<Size N, class T1, class T2>
-    static inline void peq(defaulttype::RigidCoord<N,T1>& dest, const defaulttype::Vec<N,T2>& src)
-    {
-        dest.getCenter() += src;
-    }
+    SOFA_ATTRIBUTE_DISABLED__HELPER_PEQ()
+    static inline void peq(defaulttype::RigidCoord<N, T1>& dest, const defaulttype::Vec<N, T2>& src) = delete;
+
+#undef SOFA_ATTRIBUTE_DISABLED__HELPER_EQ
+#undef SOFA_ATTRIBUTE_DISABLED__HELPER_PEQ
+
 } // namespace sofa::helper
